feature: Replace anchor.rs with hex_neighbors.
Add square and iso tilemap types to example. Swap enter and space keys.
